+++ MOVA2020/objs/dbitems/Lasku.cs
@@ -8,11 +8,11 @@ namespace MOVA2020.objs.dbitems
 {
     public class Lasku
     {
-        private int lasku_id;
+        private long lasku_id;
         private Varaus varaus;
         private double summa, alv;
 
-        public Lasku(int lasku_id, double summa, double alv, Varaus varaus)
+        public Lasku(long lasku_id, double summa, double alv, Varaus varaus)
         {
             this.lasku_id = lasku_id;
             Summa = summa;
@@ -20,7 +20,7 @@ namespace MOVA2020.objs.dbitems
             Varaus = varaus;
         }
 
-        public int Lasku_id { get => lasku_id;}
+        public long Lasku_id { get => lasku_id;}
         public double Summa { get => summa; set => summa = value; }
         public double Alv { get => alv; set => alv = value; }
         public Varaus Varaus { get => varaus; set => varaus = value; }
