+++ MOVA2020/objs/dbitems/Varaus.cs
@@ -8,12 +8,12 @@ namespace MOVA2020.objs.dbitems
 {
     public class Varaus
     {
-        private int varaus_id;
+        private long varaus_id;
         private DateTime varattu_pvm, vahvistus_pvm, varattu_alkupvm, varattu_loppupvm;
         private Asiakas asiakas;
         private Mokki mokki;
 
-        public Varaus(int varaus_id, DateTime varattu_pvm, DateTime vahvistus_pvm, DateTime varattu_alkupvm, DateTime varattu_loppupvm, Asiakas asiakas, Mokki mokki)
+        public Varaus(long varaus_id, DateTime varattu_pvm, DateTime vahvistus_pvm, DateTime varattu_alkupvm, DateTime varattu_loppupvm, Asiakas asiakas, Mokki mokki)
         {
             this.varaus_id = varaus_id;
             Varattu_pvm = varattu_pvm;
@@ -24,7 +24,7 @@ namespace MOVA2020.objs.dbitems
             Mokki = mokki;
         }
 
-        public int Varaus_id { get => varaus_id; }
+        public long Varaus_id { get => varaus_id; }
         public DateTime Varattu_pvm { get => varattu_pvm; set => varattu_pvm = value; }
         public DateTime Vahvistus_pvm { get => vahvistus_pvm; set => vahvistus_pvm = value; }
         public DateTime Varattu_alkupvm { get => varattu_alkupvm; set => varattu_alkupvm = value; }
